moved aStartWithPoint to apoc-full
